Add files via upload
These programs are used to communicate between two IP addresses. pythonClient.py sends a message to an IP address given by a user that will be received by the IP address when pythonServer.py is being run on the other computer. The program will send a message back to the client containing the number of characters and spaces and whether or not the message contained numbers.